Refactoring and fixed linting errors
