+++ frontend/src/models/assets/transport/Transport.ts
@@ -1,4 +1,6 @@
-export class Transport implements Components.Schemas.TransportDto{
+import { TransportCostProfile } from "./TransportCostProfile"
+
+export class Transport implements Components.Schemas.TransportDto {
     id?: string | undefined
     name?: string | null
     projectId?: string | undefined
@@ -19,26 +21,5 @@ export class Transport implements Components.Schemas.TransportDto{
 
     static fromJSON(data: Components.Schemas.SurfDto): Transport {
         return new Transport(data)
-    }
-
-}
-
-export class TransportCostProfile implements Components.Schemas.TransportCostProfileDto{
-    startYear?: number | undefined
-    values?: number [] | null
-    epaVersion?: string | null
-    currency?: Components.Schemas.Currency | undefined
-    sum?: number | undefined
-
-    constructor(data?: Components.Schemas.TransportCostProfileDto) {
-        this.startYear = data?.startYear
-        this.values = data?.values ?? []
-        this.epaVersion = data?.epaVersion ?? null
-        this.currency = data?.currency
-        this.sum = data?.sum
-    }
-
-    static fromJSON(data?: Components.Schemas.TransportCostProfileDto): TransportCostProfile {
-        return new TransportCostProfile(data)
     }
 }
